Remove superfluous blockquote start tokens and fix padding alignment
Previously the `Start(BlockQuote)` event would be output as "\n > ". Now
it's output as " > ". The issue was exacerbated by nested blockquotes,
which would add their own set of additional start tokens.

When mixing nested blockquotes and lists we ran into issues with
padding alignment. Because blockquotes add a leading space when we
reached the `End(BlockQuote)` event we would pop " > " off the padding
stack, and subsequent nested lists would be off by the leading space
introduced by the blockquote. Now we add additional alignment padding to
account for these scenarios.

For example:

```markdown
<!-- fix for superfluous blockquote start -->
> a
> > b
> c

<!-- fix for superfluous blockquote start -->
- > * foo
  >   * baz
  - > bar
```

would be output as (TODO: fix example):

```markdown
<!-- fix for superfluous blockquote start -->
 >
 > a
 > >
 > > b
 > c

<!-- fix for superfluous blockquote start -->
- > * foo
  >   * baz
  - > bar
```

Now it's output as:

```markdown
<!-- fix for superfluous blockquote start -->
 > a
 >  > b
 > c

<!-- fix for superfluous blockquote start -->
*  > * foo
   >   * baz
   *  > bar
```